[PL] TimeLoop: Move SingleProcessData in own file.
